requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
fake-bpy-module-latest==20230514
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
Werkzeug==2.3.4
